updated healthcare/dbconnect/queries.py

+++ healthcare/dbconnection/queries.py
@@ -16,61 +16,83 @@ def doctlist(data):
     except KeyError:
         spec_id = None
 
-    if spec_id == None:
+    if spec_id == None or spec_id=="":
         return Doctor.objects.all()
     
-    return Doctor.objects.filter(DoctorSpec__spec_id=spec_id)
+    doc2spec =  DoctorSpec.objects.filter(spec_id=spec_id)
+
+    result=[]
+
+    for doc in doc2spec:
+        result.append(Doctor.objects.filter(doctor_id=doc.doctor_id_id)[0])
+
+    return result
+        
     
 
 def newappoint(data):
-    doct_id = data["spec_id"]
+    doct_id = data["doct_id"]
     cust_id = data["cust_id"]
     date = data["date"]
     time = data["time"]
 
-    new_appointment = Appiontment(doctor_id=doct_id, cust_id=cust_id,date=date,time_slot=time)
+    doc = Doctor.objects.get(doctor_id=doct_id)
+    cust = Patient.objects.get(cust_id=cust_id)
+
+    new_appointment = Appiontment(doctor_id_id=doc.doctor_id, cust_id_id=cust.cust_id,date=date,time_slot=time)
 
     try:
         new_appointment.save()
-        return json.loads('{"query_success":"1"}')
-    except Exception:
-        return json.loads('{"query_success":"0"}')
+        return '[{"query_success":"1"}]'
+    except Exception as e:
+        return '[{"query_success":"0","error":'+str(e)+'}]'
 
 
 
 def changeappoint(data):
+    cust_id = data["cust_id"]
     appoint_id = data["appoint_id"]
     date = data["date"]
     time = data["time"]
 
-    alt_appointment = Appiontment.objects.get(appointment_id=appoint_id)
+    try:
+        alt_appointment = Appiontment.objects.get(appointment_id=appoint_id,cust_id=cust_id)
+    except Exception as e:
+        return '[{"query_success":"0","error":'+str(e)+'}]'
 
     alt_appointment.date=date
     alt_appointment.time_slot=time
 
     try:
         alt_appointment.save()
-        return json.loads('{"query_success":"1"}')
-    except Exception:
-        return json.loads('{"query_success":"0"}')
+        return '[{"query_success":"1"}]'
+    except Exception as e:
+        return '[{"query_success":"0","error":'+str(e)+'}]'
 
 def deleteappoint(data):
+    cust_id = data["cust_id"]
     appoint_id = data["appoint_id"]
 
-    del_appointment = Appiontment.objects.get(appointment_id=appoint_id)
+    try:
+        del_appointment = Appiontment.objects.get(appointment_id=appoint_id,cust_id=cust_id)
+    except Exception as e:
+        return '[{"query_success":"0","error":'+str(e)+'}]'
+
+    if del_appointment==None:
+        return '[{"query_success":"0","error":"No appointment was found with above detailes"}]'
 
     try:
         del_appointment.delete()
-        return json.loads('{"query_success":"1"}')
-    except Exception:
-        return json.loads('{"query_success":"0"}')
+        return '[{"query_success":"1"}]'
+    except Exception as e:
+        return '[{"query_success":"0","error":'+str(e)+'}]'
 
 def listappoint(data):
-    userhash = data["userhash"]
+    cust_id = data["cust_id"]
 
-    customer = Patient.objects.filter(userhash=userhash)
+    customer = Patient.objects.filter(cust_id=cust_id)
 
-    return Appiontment.objects.filter(cust_id=customer.cust_id,date__gte=datetime.date.today())
+    return Appiontment.objects.filter(cust_id=customer[0].cust_id,date__gte=datetime.date.today())
     
 
 
@@ -87,9 +109,9 @@ def runquery(request):
     querydata=request.body
     querydata_dict = json.loads(querydata)
 
-    func_to_exc = querydata["function"]
+    func_to_exc = querydata_dict["function"]
 
-    result = func_to_exc(querydata["data"])
+    result = map2func[func_to_exc](querydata_dict["data"])
 
     try:
         result_json = serializers.serialize('json', result)
